# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  CreateVideoQuestionBody:
    properties:
      portalId:
        type: integer
      timeToAnswer:
        type: integer
      timeToPrepare:
        type: integer
      title:
        type: string
      totalAttempt:
        type: integer
    required:
    - portalId
    - timeToAnswer
    - timeToPrepare
    - title
    - totalAttempt
    type: object
  CreateVideoQuestionResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      portalId:
        type: integer
      timeToAnswer:
        type: integer
      timeToPrepare:
        type: integer
      title:
        type: string
      totalAttempt:
        type: integer
      updatedAt:
        type: string
    type: object
...
  GetRoomContextResponse:
    properties:
      candidateId:
        type: integer
      dueDate:
        type: string
      isCodingDone:
        type: boolean
      isVideoDone:
        type: boolean
      roomId:
        type: integer
      totalCodingQuestion:
        type: integer
      totalCodingTime:
        type: integer
      totalVideoQuestion:
        type: integer
      totalVideoTime:
        type: integer
    required:
    - candidateId
    - dueDate
    - isCodingDone
    - isVideoDone
    - roomId
    - totalCodingQuestion
    - totalCodingTime
    - totalVideoQuestion
    - totalVideoTime
    type: object
  GetVideoQuestionByIdResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      portalId:
        type: integer
      timeToAnswer:
        type: integer
      timeToPrepare:
        type: integer
      title:
        type: string
      totalAttempt:
        type: integer
      updatedAt:
        type: string
    type: object
  GetVideoQuestionByPortalIdResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      portalId:
        type: integer
      timeToAnswer:
        type: integer
      timeToPrepare:
        type: integer
      title:
        type: string
      totalAttempt:
        type: integer
      updatedAt:
        type: string
    type: object
...
  MailObject:
    properties:
      dueDate:
        type: string
      name:
        type: string
      roomId:
        type: integer
      to:
        type: string
    required:
    - name
    - roomId
    - to
    type: object
...
  SendMailBody:
    properties:
      mailList:
        items:
          $ref: '#/definitions/MailObject'
        type: array
      preset:
        $ref: '#/definitions/handlers.MailPreset'
    required:
    - mailList
    - preset
    type: object
  UpdateRoomContextBody:
    properties:
      candidateId:
        type: integer
      dueDate:
        type: string
      isCodingDone:
        type: boolean
      isVideoDone:
        type: boolean
      roomId:
        type: integer
      totalCodingQuestion:
        type: integer
      totalCodingTime:
        type: integer
      totalVideoQuestion:
        type: integer
      totalVideoTime:
        type: integer
    required:
    - roomId
    type: object
  UpdateVideoQuestionBody:
    properties:
      id:
        type: integer
      portalId:
        type: integer
      timeToAnswer:
        type: integer
      timeToPrepare:
        type: integer
      title:
        type: string
      totalAttempt:
        type: integer
    required:
    - id
    type: object
...
  handlers.ErrResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      timestamp:
        type: string
    type: object
  handlers.MailPreset:
    enum:
    - invite
    - finish
    type: string
    x-enum-varnames:
    - Invite
    - Finish
...
  handlers.Response-CreateVideoQuestionResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/CreateVideoQuestionResponse'
      message:
        type: string
      timestamp:
        type: string
    type: object
...
  handlers.Response-GetRoomContextResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/GetRoomContextResponse'
      message:
        type: string
      timestamp:
        type: string
    type: object
  handlers.Response-GetVideoQuestionByIdResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/GetVideoQuestionByIdResponse'
      message:
        type: string
      timestamp:
        type: string
    type: object
...
  handlers.Response-array_GetVideoQuestionByPortalIdResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/GetVideoQuestionByPortalIdResponse'
        type: array
      message:
        type: string
      timestamp:
        type: string
    type: object
...
  handlers.Response-string:
    properties:
      code:
        type: integer
      data:
        type: string
      message:
        type: string
      timestamp:
        type: string
    type: object
info:
  contact: {}
  title: Interv API
  version: "1.0"
paths:
...
  /mail.sendMail:
    post:
      consumes:
      - application/json
      operationId: sendMail
      parameters:
      - description: Mail details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/SendMailBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
      summary: Send mail to the user
      tags:
      - mail
...
  /room.getRoomContext:
    get:
      consumes:
      - application/json
      operationId: getRoomContext
      parameters:
      - in: query
        name: roomId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response-GetRoomContextResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
      summary: Get room context
      tags:
      - room
  /room.updateRoomContext:
    post:
      consumes:
      - application/json
      operationId: updateRoomContext
      parameters:
      - description: update room context
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/UpdateRoomContextBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
      summary: Update room context
      tags:
      - room
...
  /videoQuestion.createVideoQuestion:
    post:
      consumes:
      - application/json
      operationId: createVideoQuestion
      parameters:
      - description: CreateVideoQuestionBody
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CreateVideoQuestionBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response-CreateVideoQuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
      summary: Create new video question
      tags:
      - videoQuestion
  /videoQuestion.deleteVideoQuestionById:
    post:
      consumes:
      - application/json
      operationId: deleteVideoQuestionById
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
      summary: Delete video question by id
      tags:
      - videoQuestion
  /videoQuestion.getVideoQuestionById/{id}:
    get:
      consumes:
      - application/json
      operationId: getVideoQuestionById
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response-GetVideoQuestionByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
      summary: Get video question by id
      tags:
      - videoQuestion
  /videoQuestion.getVideoQuestionByPortalId:
    get:
      consumes:
      - application/json
      operationId: getVideoQuestionByPortalId
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response-array_GetVideoQuestionByPortalIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
      summary: Get video question by portal id
      tags:
      - videoQuestion
  /videoQuestion.updateVideoQuestion:
    post:
      consumes:
      - application/json
      operationId: updateVideoQuestion
      parameters:
      - description: UpdateVideoQuestionBody
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/UpdateVideoQuestionBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response-CreateVideoQuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrResponse'
      summary: Update video question
      tags:
      - videoQuestion
...
swagger: "2.0"